Notes can be entered via UUID.

+++ Notiz-Speicher/src/main/java/de/thowl/prog3/exam/web/gui/CategoryFormController.java
@@ -3,13 +3,17 @@ package de.thowl.prog3.exam.web.gui;
 import de.thowl.prog3.exam.service.AuthService;
 import de.thowl.prog3.exam.storage.entities.Category;
 import de.thowl.prog3.exam.storage.repositories.CategoryRepository;
+import de.thowl.prog3.exam.storage.repositories.NoteRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
 /**
  * Controller for handling Category-related forms
  * Allows users to add and delete Categories
@@ -22,6 +26,7 @@ public class CategoryFormController {
     private static final Logger log = LoggerFactory.getLogger(NoteFormController.class);
 
     @Autowired private CategoryRepository categoryRepository;
+    @Autowired private NoteRepository noteRepository;
 
     /**
      * Adds a new Category for the authenticated User
@@ -53,13 +58,18 @@ public class CategoryFormController {
         return "redirect:/user/profile";
     }
     /**
-     * Deletes a category based on its ID
+     * Deletes a category based on its ID, while there is no note active with it
      *
      * @param categoryId The ID of the Category to delete
+     * @param redirectAttributes redirect error Attribute
      * @return Redirects to the Users profile page
      */
     @DeleteMapping("/user/category/{categoryId}")
-    public String deleteCategory(@PathVariable Long categoryId) {
+    public String deleteCategory(@PathVariable Long categoryId, RedirectAttributes redirectAttributes) {
+        if (noteRepository.existsByCategoryId(categoryId)) {
+            redirectAttributes.addFlashAttribute("error", "Diese Kategorie kann nicht gelöscht werden, da sie noch in Notizen verwendet wird.");
+            return "redirect:/user/profile";
+        }
         categoryRepository.deleteById(categoryId);
         log.info("Kategorie gefunden: {}", categoryId);
         return "redirect:/user/profile";
